Clean up classes.
